・チャート表示のTodoデータ取得ロジックを別ファイルに切り出し
